Wrap question about how to handle config in a loop
This commit attempts to address these two issues:

Whenever calendar-cli is run on a clean new user account lacking a
~/.config directory, configuration is lost and has to be remade. There
might be other cases for an Exception here and they are all annoying
enough to warrant a change in behaviour.

From a usability perspective, it is easy to be pressing enter a bit too
quickly to stop at the save prompt. This is especially true if merely
changing one or a couple of values while leaving the rest of them at
their old value. Reasking to require an actual answer on what to do
would be an improvement.