Communicate to tests using argument
Remove side effects of sourcing `setup_alias` and makes setting the test
command explicit. As a desired consequence the test command needs to be
given as an argument to `tests.sh`.

This commit additionally attempts to make global variables somewhat
easier to spot by prefixing them with underscores.

As an additional bonus, search for calendar-cli.py in ../bin/, as there
is where it typically needs to reside.